+++ Changelog.md
@@ -1,4 +1,6 @@
 # Unreleases (YYYY-MM-DD)
+
+- Add reverse binary operators, e.g. `__radd__`.
 - ...
 
 # 1.4.1 (2023-08-25)